requirements.txt
fastapi
pydantic
uvicorn
requests
google-auth
google-cloud-bigquery
pytest